e2e(ssr): handle flattening within the groq handler
